\name{closeModel}
\alias{closeModel}
\alias{closeModel}
\title{
    closeModel
}
\description{
    This connects to Machina web service, and closes the open model.
}
\usage{
    closeModel(verbose = FALSE)
}
\arguments{
    \item{verbose}{
        Not required, whether to produce verbose output (default = FALSE)
	}
}
\details{
    This connects to Machina web service, and closes the open model.
}
\value{
    This function closes the open model.
}
\author{
    Michael Gursky
}
\examples{
   # closeModel()
}
\keyword{Machina model}