\name{drawRows}
\alias{drawRows}
\alias{drawRows}
\title{
    drawRows
}
\description{
    This adds one or multiple rows to a chart window at https://account.machi.na.
}
\usage{
    drawRows(rowsArray,startDate = NULL,endDate = NULL,chartId='', verbose = FALSE)
}
\arguments{
    \item{rowsArray}{
        Not required, whether to produce verbose output (default = FALSE)
		}
    \item{startDate}{
        Not required, start date (default = NULL)
		}
    \item{endDate}{
        Not required, end date (default = NULL)
		}
    \item{chartId}{
        Not required, chart ID (default = '')
		}
    \item{verbose}{
        Not required, whether to produce verbose output (default = FALSE)
		}
}
\details{
    This adds one or multiple rows to a chart window at https://account.machi.na.
}
\value{
    This adds one or multiple rows to a chart window at https://account.machi.na.
}
\author{
    Akram Kamal
}
\examples{
  #  chart <- drawRows(c(2,3))
}
\keyword{time series chart Machina}