\name{endSession}
\alias{endSession}
\alias{endSession}
\title{
    endSession
}
\description{
    This connects to Machina web service, and ends your Machina session.
}
\usage{
    endSession(verbose = FALSE)
}
\arguments{
    \item{verbose}{
        Not required, whether to produce verbose output (default = FALSE)
	}
}
\details{
    This connects to Machina web service, and ends your Machina session.
}
\value{
    This function returns whether the session is successfully closed.
}
\author{
    Michael Gursky
}
\examples{
  #  endSession()
}
\keyword{Machina}