\name{getModel}
\alias{getModel}
\alias{getModel}
\title{
    getModel
}
\description{
    This connects to the Machina web service, and gets all the rows in the open model.
}
\usage{
    getModel(verbose = FALSE)
}
\arguments{
    \item{verbose}{
        Not required, whether to produce verbose output (default = FALSE)
	}
}
\details{
    This connects to Machina web service, and gets all the rows in the open model.
}
\value{
    This function and gets all the rows in the open model.
}
\author{
    Michael Gursky
}
\examples{
  #  getModel()
}
\keyword{Machina get model}