\name{getRemoteChartUrl}
\alias{getRemoteChartUrl}
\alias{getRemoteChartUrl}
\title{
    getRemoteChartUrl
}
\description{
    This opens the chart browser (https://account.machi.na/). It's not necessary to do this separately from drawRows, as drawRows will also open a charts browser.
}
\usage{
    getRemoteChartUrl(chartId='',openBrowser=TRUE)
}
\arguments{
	\item{chartId}{
	    Not required, specify a chart id (default = '')
	}
    \item{openBrowser}{
        Not required, whether to open chart browser (https://account.machi.na/) (default = TRUE)
	}
}
\details{
    This opens the chart browser (https://account.machi.na/).
}
\value{
    This opens the chart browser (https://account.machi.na/).
}
\author{
    Akram Kamal
}
\examples{
   # getRemoteChartUrl()
}
\keyword{chart model}