\name{listModels}
\alias{listModels}
\alias{LISTMODELS}
\title{
    listModels
}
\description{
    This connects to the Machina web service, and gets a list of your models.
}
\usage{
    listModels(verbose = FALSE)
}
\arguments{
    \item{verbose}{
        Not required, whether to produce verbose output (default = FALSE)
	}
}
\details{
    This connects to Machina web service, and and gets a list of your models.
}
\value{
    This function returns a list of your models.
}
\author{
    Michael Gursky
}
\examples{
  #  listModels()
}
\keyword{Machina list models}