\name{viewModel}
\alias{viewModel}
\alias{viewModel}
\title{
    viewModel
}
\description{
    This outputs the open model rows in a data frame.
}
\usage{
    viewModel(verbose = FALSE)
}
\arguments{
    \item{verbose}{
        Not required, whether to produce verbose output (default = FALSE)
	}
}
\details{
    This outputs the open model rows in a data frame.
}
\value{
    This outputs the open model rows in a data frame.
}
\author{
    Michael Gursky
}
\examples{
   # viewModel()
}
\keyword{Machina view model}